requirements.txt

fastapi
uvicorn
pytesseract
opencv-python
numpy
pillow
